moved the containerRunner and dockerRunner logic out of ..
        .. containerManager and just into the runner package for now
until better organization is needed
